backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
alembic==1.12.0
    # via
    #   flask-migrate
    #   pypn-ref-geo
    #   pypnusershub
amqp==5.1.1
    # via kombu
async-timeout==4.0.3
    # via redis
attrs==23.1.0
    # via fiona
authlib==1.2.1
    # via pypnusershub
bcrypt==4.0.1
    # via pypnusershub
billiard==3.6.4.0
    # via celery
blinker==1.6.2
    # via flask-mail
boto3==1.28.47
    # via taxhub
botocore==1.31.47
    # via
    #   boto3
    #   s3transfer
cached-property==1.5.2
    # via kombu
cairocffi==1.6.1
    # via
    #   cairosvg
    #   weasyprint
cairosvg==2.7.1
    # via weasyprint
celery[redis]==5.2.7
    # via -r requirements-common.in
certifi==2023.7.22
    # via
    #   fiona
    #   requests
cffi==1.15.1
    # via
    #   cairocffi
    #   cryptography
    #   weasyprint
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   -r requirements-common.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   fiona
    #   flask
    #   taxhub
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   fiona
click-repl==0.3.0
    # via celery
cligj==0.7.2
    # via fiona
cryptography==41.0.3
    # via authlib
cssselect2==0.7.0
    # via
    #   cairosvg
    #   weasyprint
defusedxml==0.7.1
    # via cairosvg
fiona==1.8.22
    # via
    #   -r requirements-common.in
    #   utils-flask-sqlalchemy-geo
flask==2.2.5
    # via
    #   -r requirements-common.in
    #   flask-admin
    #   flask-cors
    #   flask-mail
    #   flask-marshmallow
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-weasyprint
    #   flask-wtf
    #   pypn-habref-api
    #   pypn-ref-geo
    #   pypnnomenclature
    #   pypnusershub
    #   taxhub
    #   utils-flask-sqlalchemy
flask-admin==1.6.1
    # via
    #   -r requirements-common.in
    #   pypnnomenclature
flask-cors==4.0.0
    # via
    #   -r requirements-common.in
    #   taxhub
flask-mail==0.9.1
    # via -r requirements-common.in
flask-marshmallow==0.14.0
    # via
    #   -r requirements-common.in
    #   pypn-habref-api
    #   pypnnomenclature
    #   pypnusershub
flask-migrate==4.0.5
    # via
    #   -r requirements-common.in
    #   pypn-habref-api
    #   pypnnomenclature
    #   taxhub
    #   utils-flask-sqlalchemy
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements-common.in
    #   flask-migrate
    #   pypn-habref-api
    #   pypn-ref-geo
    #   pypnnomenclature
    #   pypnusershub
    #   taxhub
    #   utils-flask-sqlalchemy
flask-weasyprint==1.0.0
    # via -r requirements-common.in
flask-wtf==1.1.1
    # via -r requirements-common.in
geoalchemy2==0.11.1
    # via utils-flask-sqlalchemy-geo
geojson==3.0.1
    # via
    #   -r requirements-common.in
    #   utils-flask-sqlalchemy-geo
gunicorn==21.2.0
    # via
    #   -r requirements-common.in
    #   taxhub
html5lib==1.1
    # via weasyprint
idna==3.4
    # via requests
importlib-metadata==4.13.0 ; python_version < "3.10"
    # via
    #   -r requirements-common.in
    #   alembic
    #   attrs
    #   celery
    #   click
    #   flask
    #   gunicorn
    #   kombu
    #   mako
    #   munch
    #   redis
importlib-resources==5.12.0
    # via alembic
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.2.4
    # via celery
lxml==4.9.3
    # via -r requirements-common.in
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
marshmallow==3.19.0
    # via
    #   -r requirements-common.in
    #   flask-marshmallow
    #   marshmallow-geojson
    #   marshmallow-sqlalchemy
    #   pypn-habref-api
    #   utils-flask-sqlalchemy
marshmallow-geojson==0.4.0
    # via utils-flask-sqlalchemy-geo
marshmallow-sqlalchemy==0.28.2
    # via
    #   -r requirements-common.in
    #   pypn-habref-api
    #   pypnnomenclature
    #   pypnusershub
    #   taxhub
    #   utils-flask-sqlalchemy-geo
munch==4.0.0
    # via fiona
packaging==23.1
    # via
    #   -r requirements-common.in
    #   geoalchemy2
    #   gunicorn
    #   marshmallow
    #   marshmallow-sqlalchemy
pillow==9.5.0
    # via
    #   -r requirements-common.in
    #   cairosvg
    #   taxhub
    #   weasyprint
prompt-toolkit==3.0.39
    # via click-repl
psycopg2==2.9.7
    # via
    #   -r requirements-common.in
    #   pypn-habref-api
    #   pypn-ref-geo
    #   pypnnomenclature
    #   pypnusershub
    #   taxhub
pycparser==2.21
    # via cffi
pyphen==0.14.0
    # via weasyprint
pypn-habref-api==0.3.2
    # via -r requirements-dependencies.in
pypn-ref-geo==1.4.0
    # via
    #   -r requirements-dependencies.in
    #   taxhub
pypnnomenclature==1.5.4
    # via -r requirements-dependencies.in
pypnusershub==1.6.10
    # via
    #   -r requirements-dependencies.in
    #   pypnnomenclature
    #   taxhub
python-dateutil==2.8.2
    # via
    #   -r requirements-common.in
    #   botocore
    #   utils-flask-sqlalchemy
python-dotenv==0.21.1
    # via
    #   pypn-habref-api
    #   pypn-ref-geo
    #   pypnnomenclature
    #   taxhub
pytz==2023.3.post1
    # via celery
redis==5.0.0
    # via celery
requests==2.31.0
    # via
    #   pypn-habref-api
    #   pypnusershub
s3transfer==0.6.2
    # via boto3
shapely==1.8.5.post1
    # via
    #   -r requirements-common.in
    #   pypnnomenclature
    #   utils-flask-sqlalchemy-geo
six==1.16.0
    # via
    #   fiona
    #   flask-marshmallow
    #   html5lib
    #   python-dateutil
sqlalchemy==1.3.24
    # via
    #   -r requirements-common.in
    #   alembic
    #   flask-sqlalchemy
    #   geoalchemy2
    #   marshmallow-sqlalchemy
    #   pypn-habref-api
    #   pypn-ref-geo
    #   pypnnomenclature
    #   pypnusershub
    #   taxhub
    #   utils-flask-sqlalchemy
    #   utils-flask-sqlalchemy-geo
    #   wtforms-sqlalchemy
taxhub==1.12.1
    # via
    #   -r requirements-dependencies.in
    #   pypnnomenclature
tinycss2==1.2.1
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
toml==0.10.2
    # via -r requirements-common.in
typing-extensions==4.7.1
    # via
    #   alembic
    #   async-timeout
    #   importlib-metadata
    #   redis
urllib3==1.26.16
    # via
    #   botocore
    #   requests
    #   taxhub
utils-flask-sqlalchemy==0.3.6
    # via
    #   -r requirements-dependencies.in
    #   pypn-habref-api
    #   pypn-ref-geo
    #   pypnnomenclature
    #   pypnusershub
    #   taxhub
    #   utils-flask-sqlalchemy-geo
utils-flask-sqlalchemy-geo==0.2.8
    # via
    #   -r requirements-dependencies.in
    #   pypn-ref-geo
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
weasyprint==52.5
    # via
    #   -r requirements-common.in
    #   flask-weasyprint
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
werkzeug==2.2.3
    # via flask
wtforms==3.0.1
    # via
    #   -r requirements-common.in
    #   flask-admin
    #   flask-wtf
    #   wtforms-sqlalchemy
wtforms-sqlalchemy==0.3
    # via -r requirements-common.in
xmltodict==0.13.0
    # via -r requirements-common.in
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
